rename: components common 파일 추가
