+++ src/components/feature/login/loginHeader.tsx
@@ -2,7 +2,7 @@
 import { useRouter } from 'next/navigation'
 import React from 'react'
 
-import Header from '@/components/header'
+import Header from '@/components/common/header'
 import { Button } from '@/components/ui/button'
 import { CloseIcon } from '@/lib/icons'
